backend/requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-cloud-aiplatform
pydantic
google-cloud-secret-manager
